Trial global var controlled version of shared flush logic
This is a very rough trial of an idea which proposes that the only thing
which needs to be shared between the coordinating components is
knowledge of which one of them is supposed to be operating.

I was previously trying to get the two components to share the flush
logic timing hook, but really the only thing that needs to be assured is
that when one of them is executing its flushing, the other one is
preventing or cancelling theirs (if relevant).

All that's required to do that is a global state object which identifies
which one is supposed to be operating, and an implementation pattern
whereby all of the flush calling useEffects are aware of that variable
and will prevent / cancel their actions if it is not set in their favour
/ changes to be set not in their favour.

There are a lot of variable names etc here which need to be improved and
the design can be streamlined I'm sure, but it does seem to prove the
point.

There are several TODO's included in the comments which direct some
ideas I've had.
