+++ packages/harpguru-core/src/components/harp-guru-page/harp-guru-page.tsx
@@ -6,7 +6,7 @@ import { View, StyleSheet } from 'react-native'
 import React from 'react'
 import type { ReactElement } from 'react'
 
-import { useFlushBufferedActivityToggles } from '../toggle-buffer-flusher'
+import { ToggleBufferFlusher } from '../toggle-buffer-flusher'
 import { ScaleNotification } from '../scale-notification'
 import { QuizQuestionDisplay } from '../quiz-question-display'
 import { NextPageButton } from '../next-page-button'
@@ -50,8 +50,6 @@ export const HarpGuruPage = ({
     },
   })
 
-  const flushOverrides = useFlushBufferedActivityToggles()
-
   return (
     <PanGestureHandler
       activeOffsetX={[swipeThreshold * -1, swipeThreshold]}
@@ -94,10 +92,8 @@ export const HarpGuruPage = ({
             pageOnDisplay.setValue(nextPageNumberMap[thisPage])
           }
         />
-        <QuizQuestionDisplay
-          isScreenFree={menuState === MenuStates.NoMenu}
-          flushOverrides={flushOverrides}
-        />
+        <ToggleBufferFlusher />
+        <QuizQuestionDisplay isScreenFree={menuState === MenuStates.NoMenu} />
         <ScaleNotification />
       </View>
     </PanGestureHandler>
