+++ packages/harpguru-core/src/components/harp-guru/utils/get-initial-global-state/get-initial-global-state.ts
@@ -62,6 +62,7 @@ export const getInitialGlobalState = (pageNumber: PageNumber): GlobalState => {
     activeDisplayMode: initialDisplayMode,
     lockedCovariant: initialLockedCovariant,
     bufferedActivityToggles: thisPozitionDegrees,
+    isOverridden: false,
   }
 
   return state
