+++ packages/harpguru-core/src/components/quiz-question-display/quiz-question-display.tsx
@@ -9,21 +9,14 @@ import { getSizes } from '../../styles'
 
 import { useQuizQuestionCycle } from './hooks'
 
-type FlushOverrides = [(arg0: boolean) => void, (arg0: false | number) => void]
-
 type QuizQuestionDisplayProps = {
   readonly isScreenFree: boolean
-  readonly flushOverrides: FlushOverrides
 }
 
 export const QuizQuestionDisplay = ({
   isScreenFree,
-  flushOverrides,
 }: QuizQuestionDisplayProps): React.ReactElement => {
-  const [quizQuestion, shouldDisplay] = useQuizQuestionCycle(
-    isScreenFree,
-    flushOverrides
-  )
+  const [quizQuestion, shouldDisplay] = useQuizQuestionCycle(isScreenFree)
   const [activeExperienceMode] = useGlobal('activeExperienceMode')
 
   const toneTuples = getRenderableToneTuples(quizQuestion)
