+++ packages/harpguru-core/src/components/toggle-buffer-flusher/index.ts
@@ -1,2 +1,1 @@
 export { ToggleBufferFlusher } from './toggle-buffer-flusher'
-export { useFlushBufferedActivityToggles } from './hooks'
