+++ packages/harpguru-core/src/global.d.ts
@@ -12,5 +12,6 @@ declare module 'reactn/default' {
     activeDisplayMode: DisplayModes
     lockedCovariant: CovariantMembers
     bufferedActivityToggles: ReadonlyArray<DegreeIds>
+    isOverridden: boolean
   }
 }
