+++ packages/harpguru-core/src/hooks/index.ts
@@ -2,3 +2,4 @@ export { useMenuAnimationValues } from './use-menu-animation-values'
 export { useNudgeDisplayMode } from './use-nudge-display-mode'
 export { usePrevious } from './use-previous'
 export { useScaleAndCallbackOnTap } from './use-scale-and-callback-on-tap'
+export { useFlushBufferedActivityToggles } from './use-flush-buffered-activity-toggles'
